BlinzEngine can now have 32,768 Zones instead of only 127.

+++ src/main/java/org/blinz/world/ZoneData.java
@@ -35,7 +35,7 @@ class ZoneData {
     long zoneTime = 0;
     long zoneCycles = 0;
     final UserListenerCatalog userListeners = new UserListenerCatalog();
-    private byte id;
+    private short id;
     private String zoneName;
     private boolean paused = false;
 
@@ -43,7 +43,7 @@ class ZoneData {
      * Initializes the ZoneData according to what it already stores.
      * @param zoneID
      */
-    final void init(final byte zoneID) {
+    final void init(final short zoneID) {
         for (int i = 0; i < sectors.length; i++) {
             for (int n = 0; n < sectors[i].length; n++) {
                 sectors[i][n] = new Sector(i * sectorWidth,
